requirements.txt
boto3==1.26.38
Django==4.1.3
django_rest_framework_mongoengine==3.4.1
djangorestframework==3.14.0
mongoengine==0.24.2
